Move around validation, distribute non-aleph functions.
