+++ aleph/model/validation/network.json
